+++ src/components/BottomNavigationBar/BottomNavigationBar.js
@@ -1,13 +1,13 @@
 import React from 'react';
 import styles from './BottomNavigationBar.module.scss';
-import { ReactComponent as HomeIcon } from '../../assets/outlined/home.svg';
-import { ReactComponent as HomeIconActive } from '../../assets/filled/home-filled.svg';
-import { ReactComponent as DiscoverIcon } from '../../assets/outlined/discover.svg';
-import { ReactComponent as DiscoverIconActive } from '../../assets/filled/discover-filled.svg';
-import { ReactComponent as ChatIcon } from '../../assets/outlined/chat.svg';
-import { ReactComponent as ChatIconActive } from '../../assets/filled/chat-filled.svg';
-import { ReactComponent as ProfileIcon } from '../../assets/outlined/profile.svg';
-import { ReactComponent as ProfileIconActive } from '../../assets/filled/profile-filled.svg';
+import { ReactComponent as HomeIcon } from '../../assets/svg/home.svg';
+import { ReactComponent as HomeIconActive } from '../../assets/svg/home-filled.svg';
+import { ReactComponent as DiscoverIcon } from '../../assets/svg/discover.svg';
+import { ReactComponent as DiscoverIconActive } from '../../assets/svg/discover-filled.svg';
+import { ReactComponent as ChatIcon } from '../../assets/svg/chat.svg';
+import { ReactComponent as ChatIconActive } from '../../assets/svg/chat-filled.svg';
+import { ReactComponent as ProfileIcon } from '../../assets/svg/profile.svg';
+import { ReactComponent as ProfileIconActive } from '../../assets/svg/profile-filled.svg';
 import { NavLink, useRouteMatch } from 'react-router-dom';
 
 import { t } from '../../utils/translate';
